Extract lambda to a method and then reference it
Reactions:

Wow, using a method reference, which is a best practice for this lambda
that didn't quite fit on one line, surely brings back a lot of the
"ceremony" of the abstract inner class method definition.  Does that
lead to performance improvement because of an explicit return type?  Or
would Java infer that anyway?